+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+
+# 3.16.0
+* Add `LiabilityShift` class and `getLiabilityShift` to RiskData
+* Fixed a bug where `decision reasons` was returning an un-parsed string instead of array
+
 # 3.15.0
 * Add `AchReturnCode` to transaction object
 * Add ExchangeRateQuote API